Add files via upload
I search the data source for this assignment, which are from 'https://www.kaggle.com/datasets' and 'https://data.worldbank.org/' respectively. And then my task is to convert these data to my scheme and upload to mongodatabase for us to analyse. The python code for this task and the resources are uploaded. And the authority detail for my database is in the python file.